+++ lightx2v/common/ops/mm/mm_weight.py
@@ -2490,6 +2490,7 @@ class MMWeightTP(MMWeightTemplate):
         lora_prefix="diffusion_model.blocks",
         lora_path="",
         reduce_output=True,
+        lora_column_chunks=1,
     ):
         super().__init__(
             weight_name,
@@ -2507,7 +2508,9 @@ class MMWeightTP(MMWeightTemplate):
         self.tp_size = tp_size
         self.split_dim = split_dim  # "col" for column split, "row" for row split
         self.reduce_output = reduce_output
+        self.lora_column_chunks = lora_column_chunks
         assert split_dim in ["col", "row"], f"split_dim must be 'col' or 'row', got {split_dim}"
+        assert lora_column_chunks >= 1, f"lora_column_chunks must be positive, got {lora_column_chunks}"
 
         self._mm = MM_WEIGHT_REGISTER.get(mm_type, MMWeight)(
             weight_name=weight_name,
@@ -2521,6 +2524,50 @@ class MMWeightTP(MMWeightTemplate):
             lora_path=lora_path,
         )
         self._row_split_bias = None
+
+    def _local_lora_weights(self, weight_dict):
+        down_name = self._mm.lora_down_name
+        if down_name not in weight_dict:
+            return {}
+
+        up_name = self._mm.lora_up_name
+        if up_name not in weight_dict:
+            raise KeyError(f"LoRA is missing the up tensor paired with {down_name}")
+
+        lora_down = weight_dict[down_name]
+        lora_up = weight_dict[up_name]
+        if self.tp_size > 1:
+            if self.split_dim == "row":
+                if lora_down.shape[1] % self.tp_size:
+                    raise ValueError(f"Cannot row-shard {down_name} shape {tuple(lora_down.shape)} across TP size {self.tp_size}")
+                lora_down = torch.chunk(lora_down, self.tp_size, dim=1)[self.tp_rank].contiguous()
+            else:
+                if lora_up.shape[0] % self.lora_column_chunks:
+                    raise ValueError(f"Cannot split {up_name} shape {tuple(lora_up.shape)} into {self.lora_column_chunks} fused chunks")
+                chunks = lora_up.chunk(self.lora_column_chunks, dim=0)
+                if chunks[0].shape[0] % self.tp_size:
+                    raise ValueError(f"Cannot column-shard {up_name} shape {tuple(lora_up.shape)} across TP size {self.tp_size}")
+                lora_up = torch.cat([torch.chunk(chunk, self.tp_size, dim=0)[self.tp_rank] for chunk in chunks], dim=0).contiguous()
+
+        local_weights = {down_name: lora_down, up_name: lora_up}
+        alpha_name = self._mm.lora_alpha_name
+        if alpha_name in weight_dict:
+            local_weights[alpha_name] = weight_dict[alpha_name]
+        return local_weights
+
+    def register_lora(self, weight_dict, lora_strength=1):
+        self._mm.register_lora(self._local_lora_weights(weight_dict), lora_strength)
+
+    def update_lora(self, weight_dict, lora_strength=1):
+        self._mm.update_lora(self._local_lora_weights(weight_dict), lora_strength)
+
+    def remove_lora(self):
+        self._mm.remove_lora()
+
+    def set_config(self, config=None):
+        config = {} if config is None else config
+        self.config = config
+        self._mm.set_config(config)
 
     def _extract_row_split_bias(self, clone=False):
         if self.split_dim != "row":
@@ -2585,6 +2632,11 @@ class MMWeightTP(MMWeightTemplate):
                 output = output + self._row_split_bias
 
         return output
+
+
+def unwrap_tp_weight(module):
+    """Return the concrete tensor-owning MM implementation from a TP wrapper."""
+    return module._mm if isinstance(module, MMWeightTP) else module
 
 
 @MM_WEIGHT_REGISTER("fp8-intel-xpu")
