+++ lightx2v/models/networks/minimax_h3/weights/pre_weights.py
@@ -1,7 +1,6 @@
 import torch.distributed as dist
 
 from lightx2v.common.modules.weight_module import WeightModule, WeightModuleList
-from lightx2v.models.networks.minimax_h3.weights.tensor_parallel import build_minimax_h3_tp_linear
 from lightx2v.utils.registry_factory import ATTN_WEIGHT_REGISTER, MM_WEIGHT_REGISTER, RMS_WEIGHT_REGISTER
 
 
@@ -10,8 +9,8 @@ def _linear(name, bias=False, force_fp32=False, config=None, tp_split=None):
     lora_kwargs = {"lora_prefix": "token_refiner"} if name.startswith("token_refiner.") else {}
     if config is not None and config.get("tensor_parallel", False) and tp_split is not None:
         tp_group = config["device_mesh"].get_group(mesh_dim="tensor_p")
-        return build_minimax_h3_tp_linear(
-            tp_mm_type=config.get("tp_mm_type", "TensorParallel"),
+        tp_mm_type = config.get("tp_mm_type", "TensorParallel")
+        return MM_WEIGHT_REGISTER[tp_mm_type](
             weight_name=f"{name}.weight",
             bias_name=f"{name}.bias" if bias else None,
             mm_type=kind,
@@ -19,6 +18,7 @@ def _linear(name, bias=False, force_fp32=False, config=None, tp_split=None):
             tp_rank=dist.get_rank(tp_group),
             tp_size=dist.get_world_size(tp_group),
             split_dim=tp_split,
+            lora_column_chunks=2 if ".ff.net.0.proj" in name else 1,
             **lora_kwargs,
         )
     return MM_WEIGHT_REGISTER[kind](f"{name}.weight", f"{name}.bias" if bias else None, **lora_kwargs)
