+++ lightx2v/models/networks/minimax_h3/weights/transformer_weights.py
@@ -3,7 +3,6 @@ import torch.distributed as dist
 
 from lightx2v.common.modules.weight_module import WeightModule, WeightModuleList
 from lightx2v.models.networks.minimax_h3.infer.triton_ops import MiniMaxH3TritonRope  # noqa: F401
-from lightx2v.models.networks.minimax_h3.weights.tensor_parallel import build_minimax_h3_tp_linear
 from lightx2v.utils.registry_factory import ATTN_WEIGHT_REGISTER, MM_WEIGHT_REGISTER, RMS_WEIGHT_REGISTER, ROPE_REGISTER
 
 
@@ -11,8 +10,8 @@ def _linear(config, name, bias=False, create_cuda_buffer=False, tp_split=None):
     lora_prefix = "transformer_blocks"
     if config.get("tensor_parallel", False) and tp_split is not None:
         tp_group = config["device_mesh"].get_group(mesh_dim="tensor_p")
-        return build_minimax_h3_tp_linear(
-            tp_mm_type=config.get("tp_mm_type", "TensorParallel"),
+        tp_mm_type = config.get("tp_mm_type", "TensorParallel")
+        return MM_WEIGHT_REGISTER[tp_mm_type](
             weight_name=f"{name}.weight",
             bias_name=f"{name}.bias" if bias else None,
             mm_type=config.get("dit_quant_scheme", "Default"),
@@ -20,6 +19,7 @@ def _linear(config, name, bias=False, create_cuda_buffer=False, tp_split=None):
             tp_rank=dist.get_rank(tp_group),
             tp_size=dist.get_world_size(tp_group),
             split_dim=tp_split,
+            lora_column_chunks=2 if ".ff.net.0.proj" in name else 1,
             create_cuda_buffer=create_cuda_buffer,
             lora_prefix=lora_prefix,
         )
